Did some refactoring but still have issues
Test Case 1:
Expected: Press buttons and buttons respond with a result that matches their label.
Actual: It works!

Test Case 2:
Expected: Display should display each number pressed next to each other as the buttons are being pressed.
Actual: "0Undefined" appears.
Note: How do I capture a return from the "element" function from the event listener so that I can use the results from it do the calculations?

Test Case 3:
Expected: Buttons need to be in the shape of a square.
Actual: Buttons are rectangular.
Note: I tried to play with using flexbox, but found that the grid was the only way I was able to get the buttons to form in the way they are.  But I have no way to get the buttons to be in the position I want them to be in.   So, for example, if I remove the square root √ or the "+/-" (±), all the buttons will shift out of place.